Fixing message string format for manual test

+++ spirit-sensor-gateway/communication-protocol-strategy/KvaserCanProtocolStrategy.cpp
@@ -44,8 +44,8 @@ AWLMessage KvaserCanProtocolStrategy::convertCanMessageToAwlMessage(CanMessage c
     AWLMessage awlMessage{};
 
     awlMessage.messageID = static_cast<uint64_t>(canMessage.id);
-    awlMessage.messageLength = static_cast<uint8_t>(canMessage.length);
-    awlMessage.messageFlags = static_cast<uint8_t>(canMessage.flags);
+    awlMessage.messageLength = static_cast<uint64_t>(canMessage.length);
+    awlMessage.messageFlags = static_cast<uint64_t>(canMessage.flags);
     awlMessage.messageTimestamp = static_cast<uint64_t>(canMessage.timestamp);
 
     for (int dataIndex = 0; dataIndex < canMessage.length; ++dataIndex) {
