+++ spirit-sensor-gateway/domain/Message.h
@@ -19,8 +19,8 @@
 struct AWLMessage {
     uint64_t messageID;
     uint64_t messageTimestamp;
-    uint8_t messageFlags;
-    uint8_t messageLength;
+    uint64_t messageFlags;
+    uint64_t messageLength;
     unsigned char messageData[MESSAGE_DATA_LENGTH];
 };
